change coordinates type from double to float

+++ PropertyManagement/app/src/main/java/com/example/property_management/api/FirebaseFunctionsHelper.java
@@ -42,8 +42,8 @@ public class FirebaseFunctionsHelper {
                         (int) result.get("bathroom_num"),
                         (int) result.get("parking_num"),
                         (String) result.get("address"),
-                        (Double) result.get("lat"),
-                        (Double) result.get("lng")
+                        ((Double) result.get("lat")).floatValue(),
+                        ((Double) result.get("lng")).floatValue()
                     );
                 }
             });
